Split `spi` module into several sub-modules
This follows the example set by various other modules.
